requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.27.0
google-api-python-client==2.114.0
